infra/resume_uploader/requirements.txt
flask==3.0.0
google-cloud-pubsub==2.21.1
google-api-python-client==2.108.0
google-auth==2.25.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

